Broke it and then fixed it, except this time memory is better managed.

+++ CCTRI/histogramOperations.h
@@ -8,6 +8,6 @@ using std::vector;
 #define FUNCTIONS_H_INCLUDED
 vector<double> launder(vector<int> histPoints, double min, double max, int length);
 
-vector<int> binCounts(vector<double> data, double min, double max, double binWidth, int length);
+vector<long int> binCounts(vector<double> data, double min, double max, double binWidth, int length);
 
 #endif
